addpost 수정
